Support for producing transparent GIFs. Only when using GifWriteFrame8().
This also fixes passing frames containing the transparent index to
GifWriteFrame8(), which previously resulted in overdraw instead of using
that palette color. Now that color is remapped to the nearest match instead
(which is still not great...)

Transparency disables delta coding.

An obvious extension would be to specify a RGB triple that is replaced
with transparency when using GifWriteFrame().